refactor(notification): mv from pages to components
