+++ src/components/Templates/Notifications/components/Dropdown.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { cn } from '../../../utils/cn';
+import { cn } from '../../../../utils/cn';
 
 export type DropdownRootProps = React.HTMLAttributes<HTMLDivElement>;
 export type DropdownHeaderProps = React.HTMLAttributes<HTMLDivElement>;
